pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Project #2"
authors = [
	"Artem Bulavin <bulavin_artem@inbox.ru>"
]
readme = "README.md"
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0"

[tool.poetry.scripts]
gendiff = 'gendiff.scripts.gendiff:main'

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.5"
pytest = "^8.3.4"

[tool.ruff]
line-length = 80
preview = true
select = ["E", "F", "I", "C90"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
